requirements.txt
# django stuff
django-plotly-dash
dpd_static_support
dash-bootstrap-components
django-bootstrap4
dash_mantine_components
hypothesis
dash-iconify

# omero stuff
microscopemetrics_schema @ git+https://github.com/juliomateoslangerak/microscopemetrics-schema.git
microscopemetrics @ git+https://github.com/juliomateoslangerak/microscope-metrics.git
#zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-macos-universal2/releases/download/20240131/zeroc_ice-3.6.5-cp311-cp311-macosx_11_0_universal2.whl
#zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-linux-x86_64/releases/download/20240202/zeroc_ice-3.6.5-cp311-cp311-manylinux_2_28_x86_64.whl
omero-py==5.19.0
dash-vtk